Update panel documentation
This change fixes a thing I did wrong back when writing the file with formatting the time.

It also adds a comment to the bash example to further clarify the escape code.

I also updated the link to my website because the URL was heavily outdated (the mentioned domain has expired for at least a year now) (although I'm not sure if credits are even needed - this is not that much work so feel free to remove it if you think that would be better).